refactor: simplify app.tsx by lazy loading AppContent
